github_project/requirements.txt
requests==2.31.0
pandas==2.2.0
google-cloud-bigquery==3.40.1
google-cloud-secret-manager==2.16.4
google-auth==2.49.0
pyarrow==14.0.1
db-dtypes==1.1.1
